move detector global tests into one test file
